requirements.txt
uvicorn[standard]
aiohttp[speedups]
python-multipart
imagekitio
dnspython
httptools
aiofiles
fastapi
Pillow
motor
